+++ src/main/java/io/github/scru128/Scru128Generator.java
@@ -14,20 +14,25 @@ import java.util.Random;
  * The generator offers four different methods to generate a SCRU128 ID:
  * <table border="1">
  *   <caption>Comparison of generator functions</caption>
- *   <tr><th>Flavor</th>                       <th>Timestamp</th><th>Thread-</th><th>On big clock rewind</th></tr>
- *   <tr><td>{@link #generate}</td>            <td>Now</td>      <td>Safe</td>   <td>Rewinds state</td></tr>
- *   <tr><td>{@link #generateNoRewind}</td>    <td>Now</td>      <td>Safe</td>   <td>Returns null</td></tr>
- *   <tr><td>{@link #generateCore}</td>        <td>Argument</td> <td>Unsafe</td> <td>Rewinds state</td></tr>
- *   <tr><td>{@link #generateCoreNoRewind}</td><td>Argument</td> <td>Unsafe</td> <td>Returns null</td></tr>
+ *   <tr><th>Flavor</th>                           <th>Timestamp</th><th>Thread-</th><th>On big clock rewind</th></tr>
+ *   <tr><td>{@link #generate}</td>                <td>Now</td>      <td>Safe</td>   <td>Rewinds state</td></tr>
+ *   <tr><td>{@link #generateNoRewind}</td>        <td>Now</td>      <td>Safe</td>   <td>Returns null</td></tr>
+ *   <tr><td>{@link #generateCore(long, long)}</td><td>Argument</td> <td>Unsafe</td> <td>Rewinds state</td></tr>
+ *   <tr><td>{@link #generateCoreNoRewind}</td>    <td>Argument</td> <td>Unsafe</td> <td>Returns null</td></tr>
  * </table>
  * <p>
  * Each method returns monotonically increasing IDs unless a timestamp provided is significantly (by ten seconds or
- * more) smaller than the one embedded in the immediately preceding ID. If such a significant clock rollback is
- * detected, the standard {@link #generate} rewinds the generator state and returns a new ID based on the current
- * timestamp, whereas NoRewind variants keep the state untouched and return null. Core functions offer low-level
- * thread-unsafe primitives.
+ * more by default) smaller than the one embedded in the immediately preceding ID. If such a significant clock
+ * rollback is detected, the standard {@link #generate} rewinds the generator state and returns a new ID based on the
+ * current timestamp, whereas NoRewind variants keep the state untouched and return null. Core functions offer
+ * low-level thread-unsafe primitives.
  */
 public class Scru128Generator implements Iterable<@NotNull Scru128Id>, Iterator<@NotNull Scru128Id> {
+    /**
+     * The default timestamp rollback allowance.
+     */
+    static final long DEFAULT_ROLLBACK_ALLOWANCE = 10000L; // 10 seconds
+
     private long timestamp = 0;
     private int counterHi = 0;
     private int counterLo = 0;
@@ -71,7 +76,7 @@ public class Scru128Generator implements Iterable<@NotNull Scru128Id>, Iterator<
      * @return A new SCRU128 ID object.
      */
     public synchronized @NotNull Scru128Id generate() {
-        return generateCore(System.currentTimeMillis());
+        return generateCore(System.currentTimeMillis(), DEFAULT_ROLLBACK_ALLOWANCE);
     }
 
     /**
@@ -83,7 +88,7 @@ public class Scru128Generator implements Iterable<@NotNull Scru128Id>, Iterator<
      * @return A new SCRU128 ID object.
      */
     public synchronized @Nullable Scru128Id generateNoRewind() {
-        return generateCoreNoRewind(System.currentTimeMillis());
+        return generateCoreNoRewind(System.currentTimeMillis(), DEFAULT_ROLLBACK_ALLOWANCE);
     }
 
     /**
@@ -94,21 +99,35 @@ public class Scru128Generator implements Iterable<@NotNull Scru128Id>, Iterator<
      * Unlike {@link #generate}, this method is NOT thread-safe. The generator object should be protected from
      * concurrent accesses using a mutex or other synchronization mechanism to avoid race conditions.
      *
-     * @param timestamp A 48-bit timestamp field value.
+     * @param timestamp         A 48-bit timestamp field value.
+     * @param rollbackAllowance The amount of timestamp rollback that is considered significant. A suggested value is
+     *                          {@code 10_000} (milliseconds).
      * @return A new SCRU128 ID object.
-     * @throws IllegalArgumentException if the argument is not a 48-bit positive integer.
+     * @throws IllegalArgumentException if the timestamp is not a 48-bit positive integer.
      */
-    public @NotNull Scru128Id generateCore(long timestamp) {
-        Scru128Id value = generateCoreNoRewind(timestamp);
+    public @NotNull Scru128Id generateCore(long timestamp, long rollbackAllowance) {
+        Scru128Id value = generateCoreNoRewind(timestamp, rollbackAllowance);
         if (value == null) {
             // reset state and resume
             this.timestamp = 0;
             tsCounterHi = 0;
-            value = generateCoreNoRewind(timestamp);
+            value = generateCoreNoRewind(timestamp, rollbackAllowance);
             lastStatus = Status.CLOCK_ROLLBACK;
             assert value != null;
         }
         return value;
+    }
+
+    /**
+     * Generates a new SCRU128 ID object from the timestamp passed.
+     *
+     * @param timestamp A 48-bit timestamp field value.
+     * @return A new SCRU128 ID object.
+     * @deprecated Use {@link #generateCore(long, long)}.
+     */
+    @Deprecated
+    public @NotNull Scru128Id generateCore(long timestamp) {
+        return generateCore(timestamp, DEFAULT_ROLLBACK_ALLOWANCE);
     }
 
     /**
@@ -120,22 +139,24 @@ public class Scru128Generator implements Iterable<@NotNull Scru128Id>, Iterator<
      * Unlike {@link #generateNoRewind}, this method is NOT thread-safe. The generator object should be protected from
      * concurrent accesses using a mutex or other synchronization mechanism to avoid race conditions.
      *
-     * @param timestamp A 48-bit timestamp field value.
+     * @param timestamp         A 48-bit timestamp field value.
+     * @param rollbackAllowance The amount of timestamp rollback that is considered significant. A suggested value is
+     *                          {@code 10_000} (milliseconds).
      * @return A new SCRU128 ID object.
-     * @throws IllegalArgumentException if the argument is not a 48-bit positive integer.
+     * @throws IllegalArgumentException if the timestamp is not a 48-bit positive integer.
      */
-    public @Nullable Scru128Id generateCoreNoRewind(long timestamp) {
-        final long ROLLBACK_ALLOWANCE = 10000L; // 10 seconds
-
+    public @Nullable Scru128Id generateCoreNoRewind(long timestamp, long rollbackAllowance) {
         if (timestamp < 1 || timestamp > Scru128.MAX_TIMESTAMP) {
             throw new IllegalArgumentException("`timestamp` must be a 48-bit positive integer");
+        } else if (rollbackAllowance < 0 || rollbackAllowance > Scru128.MAX_TIMESTAMP) {
+            throw new IllegalArgumentException("`rollbackAllowance` out of reasonable range");
         }
 
         lastStatus = Status.NEW_TIMESTAMP;
         if (timestamp > this.timestamp) {
             this.timestamp = timestamp;
             counterLo = random.nextInt() & Scru128.MAX_COUNTER_LO;
-        } else if (timestamp + ROLLBACK_ALLOWANCE > this.timestamp) {
+        } else if (timestamp + rollbackAllowance > this.timestamp) {
             // go on with previous timestamp if new one is not much smaller
             counterLo++;
             lastStatus = Status.COUNTER_LO_INC;