+++ src/test/java/io/github/scru128/Scru128GeneratorTests.java
@@ -13,12 +13,12 @@ class Scru128GeneratorGenerateCoreTests {
         Scru128Generator g = new Scru128Generator();
         assertEquals(g.getLastStatus(), Scru128Generator.Status.NOT_EXECUTED);
 
-        Scru128Id prev = g.generateCore(ts);
+        Scru128Id prev = g.generateCore(ts, 10000);
         assertEquals(g.getLastStatus(), Scru128Generator.Status.NEW_TIMESTAMP);
         assertEquals(prev.getTimestamp(), ts);
 
         for (long i = 0; i < 100_000; i++) {
-            Scru128Id curr = g.generateCore(ts - Math.min(9_998, i));
+            Scru128Id curr = g.generateCore(ts - Math.min(9_998, i), 10000);
             assertTrue(g.getLastStatus() == Scru128Generator.Status.COUNTER_LO_INC ||
                     g.getLastStatus() == Scru128Generator.Status.COUNTER_HI_INC ||
                     g.getLastStatus() == Scru128Generator.Status.TIMESTAMP_INC);
@@ -35,17 +35,17 @@ class Scru128GeneratorGenerateCoreTests {
         Scru128Generator g = new Scru128Generator();
         assertEquals(g.getLastStatus(), Scru128Generator.Status.NOT_EXECUTED);
 
-        Scru128Id prev = g.generateCore(ts);
+        Scru128Id prev = g.generateCore(ts, 10000);
         assertEquals(g.getLastStatus(), Scru128Generator.Status.NEW_TIMESTAMP);
         assertEquals(prev.getTimestamp(), ts);
 
-        Scru128Id curr = g.generateCore(ts - 10_000);
+        Scru128Id curr = g.generateCore(ts - 10_000, 10000);
         assertEquals(g.getLastStatus(), Scru128Generator.Status.CLOCK_ROLLBACK);
         assertTrue(prev.compareTo(curr) > 0);
         assertEquals(curr.getTimestamp(), ts - 10_000);
 
         prev = curr;
-        curr = g.generateCore(ts - 10_001);
+        curr = g.generateCore(ts - 10_001, 10000);
         assertTrue(g.getLastStatus() == Scru128Generator.Status.COUNTER_LO_INC ||
                 g.getLastStatus() == Scru128Generator.Status.COUNTER_HI_INC ||
                 g.getLastStatus() == Scru128Generator.Status.TIMESTAMP_INC);
@@ -61,13 +61,13 @@ class Scru128GeneratorGenerateCoreNoRewindTests {
         Scru128Generator g = new Scru128Generator();
         assertEquals(g.getLastStatus(), Scru128Generator.Status.NOT_EXECUTED);
 
-        Scru128Id prev = g.generateCoreNoRewind(ts);
+        Scru128Id prev = g.generateCoreNoRewind(ts, 10000);
         assertNotNull(prev);
         assertEquals(g.getLastStatus(), Scru128Generator.Status.NEW_TIMESTAMP);
         assertEquals(prev.getTimestamp(), ts);
 
         for (long i = 0; i < 100_000; i++) {
-            Scru128Id curr = g.generateCoreNoRewind(ts - Math.min(9_998, i));
+            Scru128Id curr = g.generateCoreNoRewind(ts - Math.min(9_998, i), 10000);
             assertNotNull(curr);
             assertTrue(g.getLastStatus() == Scru128Generator.Status.COUNTER_LO_INC ||
                     g.getLastStatus() == Scru128Generator.Status.COUNTER_HI_INC ||
@@ -85,16 +85,16 @@ class Scru128GeneratorGenerateCoreNoRewindTests {
         Scru128Generator g = new Scru128Generator();
         assertEquals(g.getLastStatus(), Scru128Generator.Status.NOT_EXECUTED);
 
-        Scru128Id prev = g.generateCoreNoRewind(ts);
+        Scru128Id prev = g.generateCoreNoRewind(ts, 10000);
         assertNotNull(prev);
         assertEquals(g.getLastStatus(), Scru128Generator.Status.NEW_TIMESTAMP);
         assertEquals(prev.getTimestamp(), ts);
 
-        Scru128Id curr = g.generateCoreNoRewind(ts - 10_000);
+        Scru128Id curr = g.generateCoreNoRewind(ts - 10_000, 10000);
         assertNull(curr);
         assertEquals(g.getLastStatus(), Scru128Generator.Status.NEW_TIMESTAMP);
 
-        curr = g.generateCoreNoRewind(ts - 10_001);
+        curr = g.generateCoreNoRewind(ts - 10_001, 10000);
         assertNull(curr);
         assertEquals(g.getLastStatus(), Scru128Generator.Status.NEW_TIMESTAMP);
     }